+++ src/cards/Timeline/soccer-live-timeline-editor.js
@@ -1,7 +1,6 @@
 import { LitElement, html, css } from 'lit';
 import { t, resolveLang } from '../../i18n.js';
-import { SKIN_OPTIONS, resolveSkin } from '../../skins.js';
-import { renderSkinColorControls } from '../skin-editor.js';
+import { renderSkinControls } from '../skin-editor.js';
 
 class SoccerLiveTimelineEditor extends LitElement {
   static get properties() {
@@ -123,11 +122,8 @@ class SoccerLiveTimelineEditor extends LitElement {
         </div>
         <div>
           <label class="field-label">${this._t('editor.skin')}</label>
-          <select data-config-value="skin" @change=${this._selectChanged}>
-            ${SKIN_OPTIONS.map(([val, label]) => html`<option value="${val}" ?selected=${resolveSkin(this._config) === val}>${label}</option>`)}
-          </select>
+          ${renderSkinControls(this, this._config, (k) => (this._t ? this._t(k) : k))}
         </div>
-        ${renderSkinColorControls(this, this._config)}
         <div>
           <label class="field-label">${this._t('editor.language')}</label>
           <select data-config-value="language" @change=${this._selectChanged}>